Initial Commit
The purpose of this project was to create a web based service for the
purchase and use of redeemable wine tasting cards. There were three
other group members, along with a real-world client, who we met with on
a regular basis. We had a weekly sprint cycle, where we met with our
professor to discuss the project. My responsibilities were website
functionalities, database design, and progress reports. Customers can
purchase cards on their own via a credit card, processed by Stripe, as
well as at the winery via credit card or cash. The wineries had access
to the portion of the site that allowed for redeeming uses on the cards
and on site purchasing. Administrators had access to these features, as
well as administrative tasks such as usage reports, manipulating users,
and creating new runs of cards.